pyproject.toml
[project]
name = "django-entra-external-id"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "mozilla-django-oidc>=4.0.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
]
